ghvmctl: Add commands to check if snap is running and to stop or kill it
It may be relevant to check if a snap is still running after a while,
this may not be the case in case of crashes, so add ability to check
this.

At the same time add commands to stop or kill a snap, by using systemd
so that we don't have to deal on all the subprocesses around.